Move adjust date to shared
