+++ packages/shared/src/utils/adjust-date.ts
@@ -1,4 +1,21 @@
-import { addYears, addWeeks, addMonths, addDays, addHours, addMinutes, addSeconds, addMilliseconds } from 'date-fns';
+import {
+	addYears,
+	subWeeks,
+	subYears,
+	addWeeks,
+	subMonths,
+	addMonths,
+	subDays,
+	addDays,
+	subHours,
+	addHours,
+	subMinutes,
+	addMinutes,
+	subSeconds,
+	addSeconds,
+	addMilliseconds,
+	subMilliseconds,
+} from 'date-fns';
 import { clone } from 'lodash';
 
 /**
@@ -10,12 +27,20 @@ import { clone } from 'lodash';
 export function adjustDate(date: Date, adjustment: string): Date | undefined {
 	date = clone(date);
 
+	const subtract = adjustment.startsWith('-');
+
+	if (subtract || adjustment.startsWith('+')) {
+		adjustment = adjustment.substring(1);
+	}
+
 	const match =
-		/^((?:-|\+)?\d*?\.?\d+?) *?(milliseconds?|msecs?|ms|seconds?|secs?|s|minutes?|mins?|m|hours?|hrs?|h|days?|d|weeks?|w|months?|mth|mo|years?|yrs?|y)?$/i.exec(
-			adjustment.trim()
+		/^(-?(?:\d+)?\.?\d+) *(milliseconds?|msecs?|ms|seconds?|secs?|s|minutes?|mins?|m|hours?|hrs?|h|days?|d|weeks?|w|months?|mth|mo|years?|yrs?|y)?$/i.exec(
+			adjustment
 		);
 
-	if (!match || !match[1]) return;
+	if (!match || !match[1]) {
+		return;
+	}
 
 	const amount = parseFloat(match[1]);
 	const type = (match[2] ?? 'days').toLowerCase();
@@ -26,44 +51,44 @@ export function adjustDate(date: Date, adjustment: string): Date | undefined {
 		case 'yrs':
 		case 'yr':
 		case 'y':
-			return addYears(date, amount);
+			return subtract ? subYears(date, amount) : addYears(date, amount);
 		case 'months':
 		case 'month':
 		case 'mth':
 		case 'mo':
-			return addMonths(date, amount);
+			return subtract ? subMonths(date, amount) : addMonths(date, amount);
 		case 'weeks':
 		case 'week':
 		case 'w':
-			return addWeeks(date, amount);
+			return subtract ? subWeeks(date, amount) : addWeeks(date, amount);
 		case 'days':
 		case 'day':
 		case 'd':
-			return addDays(date, amount);
+			return subtract ? subDays(date, amount) : addDays(date, amount);
 		case 'hours':
 		case 'hour':
 		case 'hrs':
 		case 'hr':
 		case 'h':
-			return addHours(date, amount);
+			return subtract ? subHours(date, amount) : addHours(date, amount);
 		case 'minutes':
 		case 'minute':
 		case 'mins':
 		case 'min':
 		case 'm':
-			return addMinutes(date, amount);
+			return subtract ? subMinutes(date, amount) : addMinutes(date, amount);
 		case 'seconds':
 		case 'second':
 		case 'secs':
 		case 'sec':
 		case 's':
-			return addSeconds(date, amount);
+			return subtract ? subSeconds(date, amount) : addSeconds(date, amount);
 		case 'milliseconds':
 		case 'millisecond':
 		case 'msecs':
 		case 'msec':
 		case 'ms':
-			return addMilliseconds(date, amount);
+			return subtract ? subMilliseconds(date, amount) : addMilliseconds(date, amount);
 		default:
 			return undefined;
 	}
